[README.md] skimage and matplotlib
Thanks for the great work!
Just got bitten by the same error described in the linked issue, using the _nomatplotlib version solved.
Maybe the source directory could be updated to use the modified _geometry.py file? If not the suggested modification may avoid others to experience the same error.

Is it correct to say in addition that the changes in the service.py file may be made from the sources directory or from the unpacked ZIP file?